requirements.txt
transformers
trl
peft
accelerate
datasets
bitsandbytes
torch
matplotlib
tqdm
google-generativeai